% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esmr_to_tidygraph.R
\name{as_tbl_graph.nesmr}
\alias{as_tbl_graph.nesmr}
\title{Convert esmr object to tidygraph}
\usage{
\method{as_tbl_graph}{nesmr}(x, ...)
}
\arguments{
\item{param1}{An esmr object to be converted.}

\item{param2}{Additional parameters for conversion (if applicable).}
}
\value{
A tidygraph object with edge attributes for total/direct effects and their standard errors.
}
\description{
Converts an esmr object to a tidygraph representation, including total and direct effects as well as standard errors for edges.
}
