% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_T.R
\name{complete_T}
\alias{complete_T}
\title{Computes direct effects and fills in constrained total effects}
\usage{
complete_T(total_effects, s)
}
\arguments{
\item{total_effects}{Total effect matrix (lower triangular).}

\item{s}{A matrix of indices with columns: row, col such that the direct effects B
of B[row, col] = 0}
}
\value{
list of direct effects B and total effects filled in
}
\description{
O(n^3)
}
