% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_to_total.R
\name{direct_to_total}
\alias{direct_to_total}
\alias{total_to_direct}
\title{Solves for total/ effects from direct effects}
\usage{
direct_to_total(B_dir, restrict_dag = TRUE)

total_to_direct(B_tot, restrict_dag = TRUE)
}
\arguments{
\item{B_dir}{Direct effect adjacency matrix}

\item{restrict_dag}{Should the function fail if B_tot is not a DAG? Otherwise, will check that spectral radius max(abs(eigenvalues)) < 1}

\item{B_tot}{Total effect adjacency matrix}
}
\value{
Total effect adjacency matrix

Direct effect adjacency matrix
}
\description{
Solves for total/ effects from direct effects
}
\examples{
X <- structure(c(0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0), dim = c(4L,4L))
print(all(total_to_direct(direct_to_total(X)) == X))
}
