% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_to_total.R
\name{direct_to_total_adj}
\alias{direct_to_total_adj}
\title{Compute direct to total effect adjacency matrix}
\usage{
direct_to_total_adj(B, remove_diag = TRUE)
}
\arguments{
\item{B}{Direct effect adjacency matrix}

\item{remove_diag}{Logical indicating whether to remove the diagonal}
}
\value{
direct_to_total_adj(B) = B + B^2 + ... + B^p
}
\description{
Does not assume that B is a valid DAG and computes `p` steps of transitions
}
\examples{
X <- structure(c(0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0), dim = c(4L,4L))
print(direct_to_total_adj(X))
}
