% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_mr.R
\name{eb_mr}
\alias{eb_mr}
\title{Empirical Shrinkage Multivariable MR}
\usage{
eb_mr(
  beta_hat_Y,
  se_Y,
  beta_hat_X,
  se_X,
  R = NULL,
  ebnm_fn = flashier:::as.ebnm.fn(prior_family = "point_normal", optmethod = "nlm"),
  max_iter = 100,
  seed = 123,
  sigma_beta = Inf,
  tol = default_precision(c(ncol(beta_hat_X) + 1, nrow(beta_hat_X))),
  pval_thresh = 1,
  lfsr_thresh = 1,
  beta_m_init = NULL,
  which_beta = NULL,
  fix_beta = FALSE,
  beta_joint = TRUE,
  est_tau = FALSE
)
}
\arguments{
\item{beta_hat_Y}{Vector of SNP-outcome associations (length p)}

\item{se_Y}{Standard errors of beta_hat_Y}

\item{beta_hat_X}{Matrix of SNP-exposure associations (p by K)}

\item{se_X}{matrix of standard errors of beta_hat_X}

\item{R}{Optional correlation matrix for overlapping samples.}

\item{ebnm_fn}{Options prior distribution family. Defaults to point-normal.}

\item{pval_thresh}{p-value threshold for beta update}
}
\description{
Empirical Shrinkage Multivariable MR
}
