% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_mr_future.R
\name{eb_mr_future}
\alias{eb_mr_future}
\title{Empirical Shrinkage Multivariable MR}
\usage{
eb_mr_future(
  beta_hat_Y,
  se_Y,
  beta_hat_X,
  se_X,
  G = NULL,
  R = NULL,
  ebnm_fn = flashier:::as.ebnm.fn(prior_family = "point_normal", optmethod = "nlm"),
  max_iter = 100,
  sigma_beta = Inf,
  tol = default_precision(c(ncol(beta_hat_X) + 1, nrow(beta_hat_X))),
  pval_thresh = 1,
  beta_m_init = NULL,
  which_beta = NULL,
  fix_beta = FALSE,
  beta_joint = TRUE,
  g_type = c("gfa", "svd"),
  svd_zthresh = 0,
  augment_G = FALSE
)
}
\arguments{
\item{beta_hat_Y}{Vector of SNP-outcome associations (length p)}

\item{se_Y}{Standard errors of beta_hat_Y}

\item{beta_hat_X}{Matrix of SNP-exposure associations (p by K)}

\item{se_X}{matrix of standard errors of beta_hat_X}

\item{G}{G matrix. If NULL, G will be estimated using the method given in g_type.}

\item{R}{Optional correlation matrix for overlapping samples.}

\item{ebnm_fn}{Options prior distribution family. Defaults to point-normal.}

\item{max_iter}{Maximum number of iterations}

\item{sigma_beta}{Optional prior variance for causal parameters}

\item{tol}{Convergence tolerance}

\item{pval_thresh}{p-value threshold (suggest 1)}

\item{beta_joint}{Use joint updates for beta (suggest TRUE)}

\item{g_type}{Method to estimate G. Suggest "gfa"}

\item{augment_G}{Augment estimated G}
}
\description{
Empirical Shrinkage Multivariable MR
}
