% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_discovery.R
\name{eip_weighted_graph}
\alias{eip_weighted_graph}
\title{This is a sum of the edge inclusion probabilities times the estimated effect sizes.}
\usage{
eip_weighted_graph(x, ...)
}
\description{
This is a sum of the edge inclusion probabilities times the estimated effect sizes.
}
