% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/wrappers.R
\name{esmr}
\alias{esmr}
\title{esmr}
\usage{
esmr(
  beta_hat_X,
  se_X,
  beta_hat_Y = NULL,
  se_Y = NULL,
  G = NULL,
  R = NULL,
  pval_thresh = NULL,
  variant_ix = NULL,
  ld_scores = NULL,
  RE = NULL,
  max_iter = 100,
  params = mvmr_default_params()
)
}
\arguments{
\item{beta_hat_X}{Matrix of SNP-exposure associations (n by p-1).}

\item{se_X}{matrix of standard errors of beta_hat_X (n by p-1).}

\item{beta_hat_Y}{Vector of SNP-outcome associations (length n)}

\item{se_Y}{Standard errors of beta_hat_Y}

\item{G}{G matrix. If NULL, G will be estimated.}

\item{R}{Nuisance correlation matrix if there is sample overlap (optional).}

\item{pval_thresh}{p-value threshold for variant selection (optional).}

\item{variant_ix}{Indices of variants to include (optional).}

\item{ld_scores}{List of ld_scores. If missing, no additional variance term will be added.}

\item{max_iter}{Maximum number of iterations.}

\item{params}{Additional parameters. See mvmr_default_params() for more information. Generally these can be left at their defaults.}
}
\description{
esmr

Empirical Shrinkage Multivariable MR
}
\author{
Jean Morrison <jvmorr@umich.edu>
}
