% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{esmr_factors}
\alias{esmr_factors}
\title{Empirical Shrinkage Multivariable MR with Factors}
\usage{
esmr_factors(
  beta_hat_X,
  se_X,
  beta_hat_Y,
  se_Y,
  beta_hat_Z,
  se_Z,
  factors_matrix,
  factors_residual_sd = NULL,
  R = NULL,
  pval_thresh = NULL,
  variant_ix = NULL,
  ld_scores = NULL,
  RE = NULL,
  max_iter = 100,
  params = factor_default_params()
)
}
\arguments{
\item{beta_hat_X}{Vector of SNP-exposure associations (length n)}

\item{se_X}{Standard errors of beta_hat_X}

\item{beta_hat_Y}{Vector of SNP-outcome associations (length n)}

\item{se_Y}{Standard errors of beta_hat_Y}

\item{beta_hat_Z}{Matrix of SNP-confounder associations (n by p-2).}

\item{se_Z}{matrix of standard errors of beta_hat_Z (n by p-2).}

\item{factors_matrix}{Factors matrix (p-2 by k).}

\item{factors_residual_sd}{Factors residual sd}

\item{R}{Nuisance correlation matrix if there is sample overlap (optional).}

\item{pval_thresh}{p-value threshold for variant selection (optional).}

\item{variant_ix}{Indices of variants to include (optional).}

\item{ld_scores}{List of ld_scores. If missing, no additional variance term will be added.}

\item{RE}{RE}

\item{max_iter}{Maximum number of iterations.}

\item{params}{Additional parameters. See factor_default_params() for more information. Generally these can be left at their defaults.}
}
\description{
Empirical Shrinkage Multivariable MR with Factors
}
