% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esmr.R
\name{esmr_workhorse}
\alias{esmr_workhorse}
\title{ESMR Workhorse Called from wrapper functions esmr, nesmr, and esmr_factors}
\usage{
esmr_workhorse(
  beta_hat_X,
  se_X,
  beta_hat_Y = NULL,
  se_Y = NULL,
  beta_hat_Z = NULL,
  se_Z = NULL,
  factors_matrix = NULL,
  factors_residual_sd = NULL,
  G = NULL,
  R = NULL,
  pval_thresh = NULL,
  variant_ix = NULL,
  ld_scores = NULL,
  RE = NULL,
  tau_init = NULL,
  fix_tau = FALSE,
  ebnm_fn = flashier::flash_ebnm(prior_family = "point_normal", optmethod = "nlm"),
  g_init = NULL,
  fix_g = FALSE,
  max_iter = 100,
  tol = "default",
  restrict_dag = TRUE,
  direct_effect_template = NULL,
  direct_effect_init = NULL,
  beta_prior_cov = NULL,
  beta_joint = TRUE,
  augment_G = TRUE,
  cond_num = 1e+10
)
}
\description{
ESMR Workhorse Called from wrapper functions esmr, nesmr, and esmr_factors
}
