% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{is_psd}
\alias{is_psd}
\title{Checks if matrix is positive semi-definite}
\usage{
is_psd(X, tol = 1e-08)
}
\arguments{
\item{X}{matrix to test}

\item{tol}{tolerance to allow for zero eigenvalues}
}
\description{
Checks if matrix is positive semi-definite
}
