% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{layered_topo_sort_coords}
\alias{layered_topo_sort_coords}
\title{Generate coordinates for layered topological sort layout}
\usage{
layered_topo_sort_coords(g, names = NULL, x_spacing = 1, y_spacing = 1)
}
\arguments{
\item{g}{A graph object (matrix, tbl_graph, or igraph)}

\item{names}{Optional character vector of node names (if g is a matrix)}
}
\description{
Generate coordinates for layered topological sort layout
}
