% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{layered_topological_sort}
\alias{layered_topological_sort}
\title{Layered topological sort}
\usage{
layered_topological_sort(g, names = NULL)
}
\arguments{
\item{adj}{A square adjacency matrix, where [i,j] = 1 means an edge from i -> j}
}
\value{
A list of layers, where each layer is a vector of node names (or indices) that can be processed in parallel
}
\description{
Layered topological sort
}
