% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_prior.R
\name{log_graph_prior}
\alias{log_graph_prior}
\title{Graph prior}
\usage{
log_graph_prior(k, n, pi_0 = 0.5, normalize = FALSE)
}
\arguments{
\item{k}{Number of edges}

\item{n}{Total number of traits}

\item{pi_0}{Parameter to penalize the number of edges. Uniform prior is 0.5, AIC is ~0.73.}
}
\value{
prior distribution on the number of edges
}
\description{
Graph prior
}
\examples{
# Prior weights for all edges on 5 traits
n <- 5
total_edges <- n*(n-1)/2
k <- seq(0, total_edges)
prior <- log_graph_prior(k, n, pi = 0.6)
}
