% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdet.R
\name{h_det}
\alias{h_det}
\alias{h_det_grad}
\alias{h_det_vec}
\alias{h_det_grad_vec}
\alias{h_det_hessian}
\alias{h_det_hessian_vec}
\title{LogDeterminant acyclicity functions}
\usage{
h_det(X, s = 1)

h_det_grad(X, s = 1)

h_det_vec(x, s = 1, d, ix = NULL)

h_det_grad_vec(x, s = 1, d, ix = NULL)

h_det_hessian(X, s = 1)

h_det_hessian_vec(x, d = sqrt(length(x)), s = 1)
}
\arguments{
\item{X}{adjacency matrix}

\item{s}{h_det parameter. Should be greater than the highest eigenvalues of W^2}

\item{x}{vector representation of the adjacency matrix}

\item{d}{number of rows/columns in the matrix}

\item{ix}{index of the non-zero elements of the matrix in the vector versions of functions}

\item{W_vec}{vector representation of a matrix (as.vector or as.numeric(W))}
}
\description{
h_det(X) and h_det_grad(X) are zero exactly when X is an adjacency matrix that is a DAG
}
\details{
From: Bello, K., Aragam, B., & Ravikumar, P. (2023)
DAGMA: Learning DAGs via M-matrices and a Log-Determinant Acyclicity Characterization (arXiv:2209.08037).
arXiv. https://doi.org/10.48550/arXiv.2209.08037
}
