% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utils.R
\name{matrix_to_str}
\alias{matrix_to_str}
\title{Coverts a matrix into a string with newlines at the rows}
\usage{
matrix_to_str(B, collapse = "")
}
\arguments{
\item{B}{A matrix}

\item{collapse}{A string to collapse the columns}
}
\description{
Coverts a matrix into a string with newlines at the rows
}
\examples{
B <- matrix(0, nrow = 4, ncol = 4)
B[lower.tri(B)] <- 1
cat(matrix_to_str(B), "\n")
}
