% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_graph_explore.R
\name{mh_graph_explore}
\alias{mh_graph_explore}
\title{Metropolis-Hastings Graph Exploration for NESMR}
\usage{
mh_graph_explore(
  dat,
  n_mvmr_res = NULL,
  pval_select = NULL,
  R = NULL,
  alpha = 5e-08,
  init_prob_threshold = 0.05,
  init_prob_method = c("pvalue", "fdr"),
  logistic_location = 4,
  logistic_scale = 1,
  mh_chain_init = list(),
  mh_chain_params = list(),
  sparse_chain = FALSE,
  dense_chain = FALSE,
  random_starts = 0,
  max_iter = 1000,
  max_nesmr_fits = 100,
  visited_graphs = list(),
  checkpoint_file = NULL,
  checkpoint_every = 0,
  temperature = FALSE,
  burnin = round(max_iter/10),
  max_heat = 5,
  verbose = FALSE,
  debug = FALSE,
  graph_edge_prior = 0.5
)
}
\arguments{
\item{dat}{A list containing GWAS summary statistics which must include `beta_hat` and `s_estimate` (matrices of effect sizes and standard errors).}

\item{n_mvmr_res}{Optional. Precomputed NESMR MVMR results. If NULL, will be computed internally. This is computed from the nesmr_complete_mvmr and is beneficial pass if graph explorations are run multiple times.}

\item{pval_select}{Optional. Matrix of p-values for variant selection. If NULL, will be computed from `dat` as 2 * pnorm(-abs( dat$beta_hat / dat$s_estimate))}

\item{R}{Optional. LD matrix or NULL. See `?esmr::esmr` for more details.}

\item{alpha}{P-value threshold for variant selection. Default is 5e-8. Ignored if `pval_select` is provided.}

\item{init_prob_threshold}{Initial probability threshold for edge inclusion based on Z-score p-values. Default is 0.05.}

\item{init_prob_method}{Method for determining initial edge inclusion threshold. Options are "pvalue", "fdr". If "pvalue", then directly use the p-values from Z-score if "fdr", then use FDR adjusted values.}

\item{logistic_location}{Location parameter for logistic scaling. Default is 5 which centers the logistic function at the midpoint of the logistic function such that Z-score of 5 maps to 0.5.}

\item{mh_chain_init}{A named list of initial adjacency matrices to start MH chains from. If empty, will start one chain from the "best_approx" graph derived from the full MVMR results.}

\item{mh_chain_params}{A list of additional parameters for the MH chains. If the chain has an entry in this list, it will use the parameters (logistic_location, logistic_scale) from this list instead of the global parameters. Some of the names of chains provided are "best_approx", "min_graph", "max_graph", "random_start_1", "random_start_2", etc.}

\item{sparse_chain}{Logical. Whether to initialize a sparse chain (single best edge). Default is FALSE.}

\item{dense_chain}{Logical. Whether to initialize a dense chain (all possible edges). Default is FALSE.}

\item{max_iter}{Maximum number of MH iterations per chain. Default is 1000.}

\item{max_nesmr_fits}{Maximum number of NESMR model fits per chain. Default is 100.}

\item{visited_graphs}{List of previously visited graphs (for warm start or checkpointing). Default is empty list.}

\item{checkpoint_file}{Optional. File path to save checkpoints. Default is NULL.}

\item{checkpoint_every}{Integer. Save checkpoint every N NESMR fits. Default is 0 (no checkpointing).}

\item{verbose}{Logical. Print progress and debug information. Default is FALSE.}

\item{logistic_scale_range}{A numeric vector of length 2 indicating the range of logistic scaling. Default c(3, 0.0001) which linearly decrease the scale from 3 to 0.0001 over [1, max_iter]. This progressively puts more weight on the ELBO difference and more on the initial Z-score MVMR estimates as the chain progresses.}
}
\value{
An object of class `nesmr_mh_graph_explore`, a list containing:
  - visited_graphs: List of all visited graphs and their ELBOs
  - norm_elbo: Normalized ELBOs for all graphs
  - mh_chain: List of graph chains (one per initialization)
  - mh_accept: List of accept/reject indicators per chain
  - mh_accept_ratio: Mean acceptance ratio per chain
  - elbo_chain: List of ELBOs per chain
  - iter: Number of iterations per chain
  - nesmr_fits: Number of NESMR fits performed
  - elbo_denom: Log-sum-exp denominator for normalization
  - mvmr_all: The full MVMR NESMR result
}
\description{
This function performs Metropolis-Hastings (MH) sampling to explore the space of possible directed acyclic graphs (DAGs) which are then fit using NESMR.
It starts from one or more initial graphs ("best", "sparse", "dense") and iteratively proposes new graphs by adding or removing edges, accepting or rejecting proposals based on the change in model evidence (ELBO) and a proposal probability.
}
\details{
Edges are proposed based on a logistic transformation of the Z-scores. The shape of the logistic function is determined from the points: (0, min_prob) and (max_Z, max_prob).
A lower max_prob will flatten the curve and make the difference between higher and lower Z-scores less. A higher max_Z will shift the curve to the right and make it more flat.
}
\examples{

library(GWASBrewer)
library(esmr)
# Generate a simple 4-node DAG and simulate using GWASBrewer
G <- matrix(
    c(
        0, 0, 0, 0,
        0.25, 0, 0, 0,
        0, 0, 0, 0,
        0, -0.15, 0.2, 0
    ),
    nrow = 4,
    byrow = 4
)
d <- ncol(G)
h2 <- 0.2
J <- 5000
N <- 20000
pi_J <- 0.1
alpha <- 5e-8

dat <- sim_mv(
    G = G,
    N = N,
    J = J,
    h2 = h2,
    pi = pi_J,
    sporadic_pleiotropy = TRUE,
    est_s = TRUE
)

Z <- dat$beta_hat / dat$s_estimate
pval_select <- 2 * pnorm(-abs(Z))
minp <- apply(pval_select, 1, min)
ix <- which(minp < alpha)
discovery_results <- mh_graph_explore(
    dat,
    pval_select = pval_select, verbose = TRUE
)
print(discovery_results)

}
