% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{nesmr}
\alias{nesmr}
\title{Network Empirical Shrinkage MR}
\usage{
nesmr(
  beta_hat_X,
  se_X,
  direct_effect_template,
  R = NULL,
  pval_thresh = NULL,
  variant_ix = NULL,
  ld_scores = NULL,
  RE = NULL,
  max_iter = 100,
  params = nesmr_default_params()
)
}
\arguments{
\item{beta_hat_X}{Matrix of SNP-trait associations (n by p)}

\item{se_X}{Matrix of standard errors of beta_hat_X}

\item{direct_effect_template}{p by p matrix with all elements equal to 0 or 1 specifying the DAG structure.}

\item{R}{Nuisance correlation matrix if there is sample overlap (optional).}

\item{pval_thresh}{p-value threshold for variant selection (optional).}

\item{variant_ix}{Indices of variants to include (optional).}

\item{ld_scores}{List of ld_scores. If missing, no additional variance term will be added.}

\item{RE}{RE}

\item{max_iter}{Maximum number of iterations.}

\item{params}{Additional parameters. See nesmr_default_params() for more information. Generally these can be left at their defaults.}
}
\description{
Network Empirical Shrinkage MR
}
