% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesmr_all_permn.R
\name{nesmr_all_permn}
\alias{nesmr_all_permn}
\title{Run NESMR on all permutation of the traits or all DAGs}
\usage{
nesmr_all_permn(
  beta_hat,
  se_beta_hat,
  B_templates = NULL,
  all_DAGs = FALSE,
  posterior_probs = TRUE,
  return_model = FALSE,
  direct_effect_init = NULL,
  variant_ix = NULL,
  alpha = 5e-08,
  log_graph_prior_pi = 0.5,
  beta_prior_cov = NULL,
  one_step_correction = FALSE,
  log_lik = FALSE,
  mc_cores = 1,
  ...
)
}
\arguments{
\item{beta_hat}{Matrix of effect estimates}

\item{se_beta_hat}{Matrix of standard errors}

\item{B_templates}{List of direct effect templates. If this is specified then only compute for these templates rather than all permutations or all DAGs.}

\item{all_DAGs}{Should all DAGs be generated? FALSE means all permutations of the traits}

\item{posterior_probs}{Should posterior probabilities be computed?}

\item{return_model}{Should the full model object be returned?}

\item{direct_effect_init}{Initial direct effect matrix for the current order of the traits}

\item{variant_ix}{Index of variants to subset}

\item{alpha}{Significance level for p-value cutoff}

\item{...}{Additional arguments to pass to `esmr`}
}
\value{
List of results
}
\description{
Run NESMR on all permutation of the traits or all DAGs
}
