% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesmr_backselect.R
\name{nesmr_backselect}
\alias{nesmr_backselect}
\title{NESMR backselect algorithm}
\usage{
nesmr_backselect(
  mod_list,
  method = c("posterior_prob", "aic", "elbo"),
  posterior_prob_cutoff = 0.005,
  aic_cutoff = 2,
  pvalue_cutoff = 0.05,
  alpha = 5e-08,
  log_graph_prior_pi = 0.5,
  ...
)
}
\arguments{
\item{mod_list}{List of initial models to search}

\item{method}{Currently only does AIC based select within `aic_cutoff` of the best current model}

\item{aic_cutoff}{AIC difference cutoff for selecting models}

\item{pvalue_cutoff}{P-value cutoff for selecting edges}

\item{alpha}{Significance level for selecting variants}

\item{...}{Additional parameters to pass into esmr}

\item{beta_hat}{Estimated beta coefficients from GWAS}

\item{se_beta_hat}{Standard errors of beta coefficients from GWAS}

\item{mod_log_liks}{Optional log-likelihoods of the model. Otherwise, look at model$log_lik and then attempt to call log_py(model)}

\item{Z_true}{Optional true Z-scores to use for selection}
}
\value{
List of models selected by NESMR backselect
}
\description{
NESMR backselect algorithm
}
\examples{
set.seed(13)
d <- 4
G <- matrix(0, nrow = d, ncol = d)
G_low <- lower.tri(G)
B_lower <- G_low + 0
G[d, 1] <- 0.1

B_true <- (G != 0) + 0

h2 <- 0.3
J <- 5000
N <- 30000
pi_J <- 0.1
alpha <- 5e-8
lambda <- qnorm(1 - alpha / 2)
dat <- GWASBrewer::sim_mv(
  G = G,
  N = N,
  J = J,
  h2 = h2,
  pi = pi_J,
  sporadic_pleiotropy = TRUE,
  est_s = TRUE
)

Ztrue <- with(dat, beta_marg/se_beta_hat)
pval_true <- 2*pnorm(-abs(Ztrue))
minp <- apply(pval_true, 1, min)
ix <- which(minp < alpha)

backselect_results <- nesmr_backselect(
  list(full_mod),
  beta_hat = dat$beta_hat,
  se_beta_hat = dat$s_estimate, Z_true = Ztrue,
  aic_cutoff = 10, pvalue_cutoff = 0.05 / sum(B_lower),
  alpha = 5e-8)

# Print the AIC of the models and indicate the true generating model
back_aic <- sapply(backselect_results, function(x) x$aic)
min_aic <- which.min(back_aic)
true_ix <- which(sapply(backselect_results, function(x) all(x$B_template == B_true)))
plot(min(back_aic) - back_aic,
    pch = ifelse(1:length(back_aic) == true_ix, 8, 1)
)
legend("topleft", legend = c("True model"),
      pch = c(8, 1))
}
