% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_differences.R
\name{plot_graph_differences}
\alias{plot_graph_differences}
\title{Plot Differences Between Two Graphs}
\usage{
plot_graph_differences(
  tg1,
  tg2,
  diff_type = c("adj", "beta"),
  effect = c("direct_effect", "total_effect"),
  pos_color = "#d62728",
  neg_color = "#1f77b4",
  scale_colors = c(`-2` = "#D7191C", `-1` = "#FDAE61", `0` = "white", `1` = "#ABDDA4",
    `2` = "#2B83BA"),
  node_order = NULL,
  x_axis_position = c("top", "bottom")
)
}
\arguments{
\item{tg1}{first graph object}

\item{tg2}{second graph object}

\item{diff_type}{Character vector specifying the type of difference to plot. Options are \code{"adj"} for adjacency matrix differences or \code{"beta"} for beta coefficient differences.}
}
\value{
A plot visualizing the differences between the two graphs.
}
\description{
This function visualizes the differences between two graphs, either by adjacency matrices or beta coefficients.
}
