% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_to_dag.R
\name{project_to_DAG}
\alias{project_to_DAG}
\title{Project a matrix to a DAG using the Frobenius norm and logdet penalty (DAGMA - Bello 2022)}
\usage{
project_to_DAG(
  X,
  s = max(max(X^2) + 0.1, 1),
  lambda = c(1, 10^-seq(1, 5), 0),
  penalty = c("L2", "L1"),
  threshold_to_DAG = FALSE,
  maxit = 100
)
}
\arguments{
\item{X}{a square matrix, not necessarily a DAG}

\item{s}{log-det penalty parameter}

\item{lambda}{vector for path finding parameter}

\item{penalty}{L2 or L1 penalty (default and more stable: L2)}

\item{threshold_to_DAG}{should the matrix be thresholded to a true DAG?
Otherwise, the matrix will have near zeros but not exact zeros.
Thresholding is done via \code{\link{threshold_DAG}} which iterativly finds the minimal threshold that ensures acyclicity.}

\item{...}{additional arguments to the `control` argument of \code{\link{optim}}}
}
\value{
a matrix that is a DAG that is closest to X by L2 norm
}
\description{
Project a matrix to a DAG using the Frobenius norm and logdet penalty (DAGMA - Bello 2022)
}
