% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_to_dag.R
\name{project_to_DAG_bootstrap}
\alias{project_to_DAG_bootstrap}
\title{Bootstrap LogDet}
\usage{
project_to_DAG_bootstrap(
  total_est,
  total_est_se,
  reps = 100,
  s = 1.1,
  maxit = 500
)
}
\arguments{
\item{total_est}{total or direct effects matrix}

\item{total_est_se}{standard errors of total or direct effects matrix}

\item{reps}{total bootstrap samples}

\item{s}{logdet parameter}

\item{...}{}
}
\value{
a list of DAGs
}
\description{
Start with an effect matrix and standard errors and sample from an (independent) MV normal distribution
and project to DAG using LogDet characterization
}
