% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate2.R
\name{sim_mv}
\alias{sim_mv}
\title{Simulate multivariate GWAS data}
\usage{
sim_mv(
  tau_xz,
  tau_yz,
  dir_xz,
  dir_yz,
  gamma,
  h2_x,
  h2_y,
  h2_z,
  N,
  J,
  pi_x,
  pi_y,
  pi_z
)
}
\arguments{
\item{dir_xz, dir_yz}{Direction of effects between each of Z_1, ..., Z_K and X and Y respectively.
A value of 1 indicates an effect from Z_k to X. A value of -1 indicates an effect from X to Z_k.}

\item{gamma}{Causal effect of X on Y.}

\item{h2_x, h2_y}{Scalars giving hertiabiltiy of each of X and Y.}

\item{h2_z}{Length K vector giving heritability of each of Z_1 to Z_K.}

\item{N}{Either a single number representing GWAS sample size for all studies or an K+2
vector giving sample size for X, Y, and then each of Z_1 to Z_K}

\item{J}{Number of variants to simulate}

\item{pi_x, pi_y}{Scalrs between 0 and 1 giving the proportion of non-zero effects for each of X and Y}

\item{pi_z}{Length K vector giving proporiton of non-zero effects for each of Z_1 to Z_K.}

\item{taux_xz, tau_yz}{Length K vectors. Effects between each of Z_1, ..., Z_K and X and Y respectively.}
}
\description{
Simulate multivariate GWAS data
}
\details{
This function simulates data for traits X, Y, and Z_1, ..., Z_k.
}
