% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{solve_diag_psd_diag}
\alias{solve_diag_psd_diag}
\title{Computes a faster version of a diagonal matrix times a square matrix
times the same diagonal matrix}
\usage{
solve_diag_psd_diag(X, s, check_psd = FALSE)
}
\arguments{
\item{X}{square matrix with dimensions d x d}

\item{s}{vector of length d}
}
\description{
Computes a faster version of a diagonal matrix times a square matrix
times the same diagonal matrix
}
