% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sum_stats}
\alias{sum_stats}
\title{Simulate summary statistics}
\usage{
sum_stats(
  nsnps,
  Nx,
  Ny,
  Nz,
  h2x,
  h2y,
  h2z,
  num_conf_z,
  num_null_z,
  nsnpx,
  nsnpy,
  nsnpz,
  taux,
  tauz,
  omegaz
)
}
\arguments{
\item{Nx, Ny, Nz}{sample size for X, Y, and Z_1, ... Z_M gwas. Assuming equal sample size
for all possible covariates for simplicity.}

\item{h2x, h2y, h2z}{Expected heritability of X, Y, and Z_1, ..., Z_M.}

\item{num_conf_z}{Number of confounders}

\item{num_null_z}{Number of covariates that don't effect either X or Y}

\item{nsnpx, nsnpy, nsnpz}{Number of effect SNPs for X, Y, and Zs}

\item{tauz}{Proportion of Y heritability explained by each confounder. Can be vector of length num_conf_z or scalar.}

\item{omegaz}{Proportion of X heritability explained by each confounder. Can be vector of length num_conf_z or scalar.}

\item{Nsnps}{Number of snps to simulate}

\item{tau}{tau = h1*(gamma^2)/h2 Proportion of Y heritability explained by X.}
}
\description{
Simulate summary statistics
}
