% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphcml_utils.R
\name{total_to_dir_projection_gcml}
\alias{total_to_dir_projection_gcml}
\title{Pulled from graph-cML Lin2023}
\usage{
total_to_dir_projection_gcml(G_tot, maxit = 100)
}
\description{
Pulled from graph-cML Lin2023
}
