% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esmr.R, R/package.R
\docType{package}
\name{esmr}
\alias{esmr}
\title{Empirical Shrinkage Multivariable MR}
\usage{
esmr(
  beta_hat_Y,
  se_Y,
  beta_hat_X,
  se_X,
  G = NULL,
  R = NULL,
  pval_thresh = 5e-08,
  ebnm_fn = flashier::flash_ebnm(prior_family = "point_normal", optmethod = "nlm"),
  max_iter = 100,
  sigma_beta = Inf,
  tol = default_precision(c(ncol(beta_hat_X) + 1, nrow(beta_hat_X))),
  direct_effect_template = NULL,
  direct_effect_init = NULL,
  beta_joint = TRUE,
  augment_G = TRUE
)
}
\arguments{
\item{beta_hat_Y}{Vector of SNP-outcome associations (length p)}

\item{se_Y}{Standard errors of beta_hat_Y}

\item{beta_hat_X}{Matrix of SNP-exposure associations (p by K)}

\item{se_X}{matrix of standard errors of beta_hat_X}

\item{G}{G matrix. If NULL, G will be estimated using the method given in g_type.}

\item{R}{Optional correlation matrix for overlapping samples.}

\item{pval_thresh}{p-value threshold (suggest 1)}

\item{ebnm_fn}{Options prior distribution family. Defaults to point-normal.}

\item{max_iter}{Maximum number of iterations}

\item{sigma_beta}{Optional prior variance for causal parameters}

\item{tol}{Convergence tolerance}

\item{beta_joint}{Use joint updates for beta (suggest TRUE)}

\item{augment_G}{Augment estimated G}

\item{g_type}{Method to estimate G. Suggest "gfa"}
}
\author{
Jean Morrison <jvmorr@umich.edu>
}
