% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esmr.R, R/package.R
\name{esmr}
\alias{esmr}
\title{Empirical Shrinkage Multivariable MR}
\usage{
esmr(
  beta_hat_X,
  se_X,
  beta_hat_Y = NULL,
  se_Y = NULL,
  G = NULL,
  R = NULL,
  pval_thresh = NULL,
  variant_ix = NULL,
  ld_scores = NULL,
  RE = NULL,
  tau_init = NULL,
  fix_tau = FALSE,
  ebnm_fn = flashier::flash_ebnm(prior_family = "point_normal", optmethod = "nlm"),
  g_init = NULL,
  fix_g = FALSE,
  max_iter = 100,
  sigma_beta = Inf,
  tol = "default",
  restrict_dag = TRUE,
  direct_effect_template = NULL,
  direct_effect_init = NULL,
  beta_joint = TRUE,
  augment_G = TRUE
)
}
\arguments{
\item{beta_hat_X}{Matrix of SNP-exposure associations (p by K)}

\item{se_X}{matrix of standard errors of beta_hat_X}

\item{beta_hat_Y}{Vector of SNP-outcome associations (length p)}

\item{se_Y}{Standard errors of beta_hat_Y}

\item{G}{G matrix. If NULL, G will be estimated using the method given in g_type.}

\item{R}{Optional correlation matrix for overlapping samples.}

\item{pval_thresh}{p-value threshold for estimation}

\item{variant_ix}{Instead of using pval_thresh, directly specify the indices of variants used for estimation.}

\item{ebnm_fn}{Options prior distribution family. Defaults to point-normal.}

\item{max_iter}{Maximum number of iterations}

\item{sigma_beta}{Optional prior variance for causal parameters}

\item{tol}{Convergence tolerance}

\item{beta_joint}{Use joint updates for beta (suggest TRUE)}

\item{augment_G}{Augment estimated G}

\item{g_type}{Method to estimate G. Suggest "gfa"}
}
\description{
Empirical Shrinkage Multivariable MR

esmr
}
\author{
Jean Morrison <jvmorr@umich.edu>
}
